--added an insanely robust version of the bye command

+++ commands/fun/bye.js
@@ -1,4 +1,5 @@
 const { Command } = require("discord.js-commando")
+const { getVoiceChannel } = require("../../helpers")
 
 module.exports = class extends Command {
   constructor (client) {
@@ -19,7 +20,12 @@ module.exports = class extends Command {
       return
     }
 
-    if (!msg.channel.id.includes("729980395151556709")) {
+    if (!msg.channel.permissionsFor(msg.guild.roles.everyone).has("VIEW_CHANNEL")) {
+      msg.channel.send("To use this command it needs to be posted in a channel that EVERYONE can see. If your doing this in a secret channel you know who you are, naughty naughty")
+      return
+    }
+
+    if (!msg.channel.name.includes("general")) {
       msg.channel.send("Post this in general please, you sneaky beaver")
       return
     }
@@ -36,11 +42,18 @@ module.exports = class extends Command {
       return
     }
 
+    const authorOfMessage = msg.guild.members.cache.get(msg.member.id)
+    const authorOriginalChannelID = getVoiceChannel(msg)
+    let peopleToBeRemoved = msg.member.voice.channel.members.map(member => member)
     const checkPeopleAreInChannel = msg.member.voice.channel.members.map(member => member.user.id)
     const confirmMsg = await msg.reply(`Is this bye wanted?`)
     confirmMsg.react("🛑")
     const filter = (reaction, user) => ["🛑"].includes(reaction.emoji.name) && !user.bot && checkPeopleAreInChannel.includes(user.id)
     const collected = await confirmMsg.awaitReactions(filter, { time: 8000, max: 1 })
+    if (confirmMsg.deleted) {
+      msg.channel.send("Rude, don't delete the bye until its finished please")
+      return
+    }
     confirmMsg.delete()
 
     const firstKey = collected.firstKey()
@@ -50,18 +63,26 @@ module.exports = class extends Command {
       return
     }
 
+    const authorCurrentChannelID = getVoiceChannel(msg)
+
+    if (authorCurrentChannelID !== authorOriginalChannelID) {
+      authorOfMessage.voice.setChannel(null)
+      msg.channel.send("It appears you've started a bye then moved to another channel. Lucille doesn't like this so i'm just gonna kick just you")
+      return
+    }
+
     const music = msg.guild.music
     music.state.queue.splice(0, music.state.queue.length)
     music.setState({ queue: music.state.queue })
 
-    const kickedMembers = msg.member.voice.channel.members.map(member => {
+    peopleToBeRemoved = peopleToBeRemoved.map(member => {
       member.voice.setChannel(null)
       return member.displayName
     })
 
-    if (kickedMembers.length) {
+    if (peopleToBeRemoved.length) {
       msg.react("👋")
-      msg.channel.send(`Goodbye ${join(kickedMembers)}`)
+      msg.channel.send(`Goodbye ${join(peopleToBeRemoved)}`)
     }
   }
 }
